+++ src/app/pages/orders-new-page/components/areas-data-form-control/components/area-data-dialog/components/map-form-control/components/map-point-select-form-control/map-point-select-form-control.component.ts
@@ -2,14 +2,12 @@
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { MapPointSelectFormControlService } from './map-point-select-form-control.service';
 import { TranslocoModule } from '@jsverse/transloco';
-import {
-    ENTRY_POINT_MARKER_OPTIONS,
-    MapPointSelectFormControlVM,
-} from './map-point-select-form-control.model';
+import { MapPointSelectFormControlVM } from './map-point-select-form-control.model';
 import { NgClass } from '@angular/common';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { Coordinates } from '@stores/location/location.model';
 import { Component, effect, forwardRef, inject, input } from '@angular/core';
+import { ENTRY_POINT_MARKER_OPTIONS } from 'src/app/shared/providers/google-maps-provider';
 
 const noop = () => {};
