+++ src/app/pages/orders-new-page/components/areas-data-form-control/components/area-data-dialog/components/map-form-control/components/map-point-select-form-control/map-point-select-form-control.model.ts
@@ -1,15 +1,3 @@
-import { InjectionToken } from '@angular/core';
-
-export const ENTRY_POINT_MARKER_OPTIONS =
-    new InjectionToken<google.maps.marker.AdvancedMarkerElementOptions>(
-        'Injection Token for entry point marker options'
-    );
-
-export const provideMockEntryPointMarkerOptions = () => ({
-    provide: ENTRY_POINT_MARKER_OPTIONS,
-    useValue: {},
-});
-
 export interface MapPointSelectFormControlVM {
     deleteButtonTextKey: string;
     addButtonTextKey: string;
