+++ src/app/pages/orders-new-page/components/areas-data-form-control/components/area-data-dialog/components/map-form-control/components/map-point-select-form-control/map-point-select-form-control.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { MapPointSelectFormControlService } from './map-point-select-form-control.service';
-import { provideMockEntryPointMarkerOptions } from './map-point-select-form-control.model';
+import { provideMockEntryPointMarkerOptions } from 'src/app/shared/providers/google-maps-provider';
 
 describe('MapPointSelectFormControlService', () => {
     let service: MapPointSelectFormControlService;
